Moved uptime class definition to it's own file
